requirements.txt
Django==1.8
whitenoise==1.0.6
psycopg2==2.5.2
dj-database-url==0.3.0
django-localflavor==1.1
django-dotenv==1.3.0
djangohelpers==0.18.1
git+git://github.com/linuxlewis/djorm-ext-pgfulltext@tags/0.9.3
django-debug-toolbar==1.3.0
django-pipeline==1.5.4
django-bower==5.0.4
django-bootstrap-form==3.1
django-bootstrap3==4.11.0
django-registration-defaults==0.4.4
django-registration-redux==1.1
